+++ src/runtime/node/async_hooks/index.ts
@@ -3,9 +3,9 @@ import type asyncHooks from "node:async_hooks";
 import { AsyncLocalStorage } from "./_async-local-storage";
 import { AsyncResource } from "./_async-resource";
 
-import * as asyncHook from "./_async_hook";
+import * as asyncHook from "./_async-hook";
 
-export * from "./_async_hook";
+export * from "./_async-hook";
 
 export default <typeof asyncHooks>{
   AsyncLocalStorage,
